requirements.txt
ujson
cchardet
urllib3
requests
lxml
scrapy
selenium
pymysql
python-dateutil
pywin32